requirements.txt
aioitertools==0.12.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bcrypt==5.0.0
bibtexparser==1.4.3
boto3==1.40.40
botocore==1.40.40
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dnspython==2.8.0
emmet-core==0.84.10
fonttools==4.60.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
invoke==2.2.0
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latexcodec==3.0.1
maggma==0.72.0
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
mongomock==4.3.0
monty==2025.3.3
mp-api==0.45.11
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.6.0
networkx==3.5
numpy==2.3.3
orjson==3.11.3
packaging==25.0
palettable==3.3.3
pandas==2.3.2
paramiko==4.0.0
pillow==11.3.0
plotly==6.3.0
pyarrow==21.0.0
pybtex==0.25.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydash==8.0.5
Pygments==2.19.2
pymatgen==2025.6.14
pymatgen-io-validation==0.1.2
pymongo==4.10.1
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
scipy==1.16.2
sentinels==1.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spglib==2.6.0
sshtunnel==0.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.5.0
wrapt==1.17.3
